+++ bn_s_mp_reverse.c
@@ -4,9 +4,9 @@
 /* SPDX-License-Identifier: Unlicense */
 
 /* reverse an array, used for radix code */
-void s_mp_reverse(unsigned char *s, int len)
+void s_mp_reverse(unsigned char *s, size_t len)
 {
-   int     ix, iy;
+   size_t   ix, iy;
    unsigned char t;
 
    ix = 0;
